corrected + completed

+++ tema1/t1p1.cpp
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-void update_u(float& u) // u is of form: 10^(-m)
+void update_u(double& u) // u is of form: 10^(-m)
 {
 	u = u/10.0;
 }
@@ -10,8 +10,8 @@ void update_u(float& u) // u is of form: 10^(-m)
 int main ()
 {
 	int m = 0;
-	float prev_u = -1;
-	float u = 1;
+	double prev_u = -1;
+	double u = 1;
 
 	// MAIN LOOP
 	while (!(1 + u == 1))
@@ -23,8 +23,8 @@ int main ()
 	}
 
 	std::cout << "Precizia masina e: " << prev_u << '\n';
-	// Andy: resulting u is 1.4013e-45
-	// Ama: u = 1e-07
+	// Andy: resulting u is 1.4013e-45 (originally)
+	// Ama: u = 1e-15
 
 	return 0;
 }
